requirements.txt
streamlit
mysql-connector-python
Pillow
openai
